refactor: prepare for v5 support
